+++ Back-end/src/main/java/com/example/backend/cart/dto/dto.txt
@@ -1,0 +1,1 @@
+클라이언트와 데이터 주고받을 때 사용하는 클래스
