+++ Back-end/src/main/java/com/example/backend/comment/entity/Comment.java
@@ -1,6 +1,8 @@
-package com.example.backend.domain;
+package com.example.backend.comment.entity;
 
 
+import com.example.backend.user.entity.User;
+import com.example.backend.board.entity.Board;
 import jakarta.persistence.*;
 import lombok.*;
 import org.hibernate.annotations.CreationTimestamp;
@@ -10,21 +12,19 @@ import java.time.LocalDateTime;
 import java.util.UUID;
 
 @Entity
-@Table(name = "board")
+@Table(name = "comment")
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
 @Builder
-public class Board {
+public class Comment {
+
     @Id
     @GeneratedValue(strategy = GenerationType.UUID)
-    @Column(name = "board_id", columnDefinition = "BINARY(16)")
-    private UUID boardId;
+    @Column(name = "comment_id", columnDefinition = "BINARY(16)")
+    private UUID commentId;
 
-    @Column(name = "title", length = 50, nullable = false)
-    private String title;
-
-    @Column(name = "content", length = 1000, nullable = false)
+    @Column(name = "content", length = 500, nullable = false)
     private String content;
 
     @CreationTimestamp
@@ -35,10 +35,12 @@ public class Board {
     @Column(name = "updated_at", nullable = false)
     private LocalDateTime updatedAt;
 
-    @Column(name = "count", nullable = false)
-    private int count;
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "board_id")
+    private Board boardId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id")
     private User userId;
+
 }
