+++ Back-end/src/main/java/com/example/backend/comment/entity/entity.txt
@@ -1,0 +1,1 @@
+DB 테이블 매핑 (Entity 클래스)
