+++ Back-end/src/main/java/com/example/backend/comment/service/service.txt
@@ -1,0 +1,1 @@
+비즈니스 로직
