+++ Back-end/src/main/java/com/example/backend/common/config/SecurityConfig.java
@@ -1,4 +1,4 @@
-package com.example.backend.config;
+package com.example.backend.common.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
