+++ Back-end/src/main/java/com/example/backend/group/repository/repository.txt
@@ -1,0 +1,1 @@
+JPA 또는 MyBatis 인터페이스
