+++ Back-end/src/main/java/com/example/backend/schedule/controller/controller.txt
@@ -1,0 +1,1 @@
+API 엔드포인트 처리
